fix: #WB2-1513, feedback PR with store + move files
